Delete useless methods and their test code
Rename c_rocks_db_crud_test to c_rocks_db_default_test because we don't use update and delete method
Delete useless methods which were copy of tendermint database interface.
Delete getter methods which were used by test
